refactor folder structure for API ussage
