+++ src/API/hooks/useCategories.ts
@@ -1,5 +1,5 @@
 import useSWR from "swr";
-import { fetchCategories } from "@/services/apiService";
+import { fetchCategories } from "@/API/tag-categories/get";
 
 export const useCategories = () => {
     const { data, error } = useSWR("/categories/all", fetchCategories, {
