+++ src/API/tag-categories/get.ts
@@ -1,0 +1,10 @@
+import axiosClient from "@/lib/axiosClient";
+
+export const fetchCategories = async () => {
+    try {
+        const response = await axiosClient.get("/categories/all");
+        return response.data ? response.data : response;
+    } catch (error) {
+        throw error;
+    }
+};
