+++ src/app/components/hompage/Promo_Section/index.tsx
@@ -1,7 +1,7 @@
 import { Carousel, PromoCard } from "@/app/components";
-import { usePromo } from "@/hooks/usePromo";
+import { usePromo } from "@/API/hooks/usePromo";
 import PromoCardSkeleton from "@/app/components/hompage/PromoCard/PromoCardSkeleton";
-import { PromoProps } from "@/types/apiTypes";
+import { PromoProps } from "@/API/types/apiTypes";
 
 const Promo = () => {
   const { promo, isLoading, isError } = usePromo();
